+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `uart::{start,stop}()` and `uart::set_baudrate()` functions
 - Rename `clock` module to `timer`
 - Make all API functions return `isize`
 - Change `debug::time()` to return `isize` and take `*mut u64`